requirements.txt
nltk == 3.5
pymorphy2 == 0.9.1
jamspell == 0.0.12
asgiref==3.3.1
Django==3.1.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.10.0
django-extensions==3.1.0
docx==0.2.4
fire==0.3.1
lxml==4.6.2
pdf2docx==0.4.6
Pillow==8.0.1
PyMuPDF==1.18.4
python-docx==0.8.10
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0